srand

+++ reduce.c
@@ -32,7 +32,7 @@ int main(int argc, char** argv) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
       printf("\nSum is %f\n", *numbers);
       double pi = *numbers/((double) world_size);
       printf("\nPi is %f\n", pi);
@@ -42,77 +42,77 @@ int main(int argc, char** argv) {
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 2) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 3) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 4) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 5) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 6) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 7) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 8) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 9) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 10) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   } else if (world_rank == 11) {
       srand(clock());
       srand(time(NULL) + world_rank);
       number = get_pi(iter_num);
       printf("\nPi in %d is %f\n", world_rank, number);
-      MPI_Reduce(&number, numbers, 1,  MPI_INT, MPI_SUM, 0, MPI_COMM_WORLD);
+      MPI_Reduce(&number, numbers, 1,  MPI_DOUBLE, MPI_SUM, 0, MPI_COMM_WORLD);
 
   }
